+++ src/main/java/chess/domain/board/BoardFactory.java
@@ -1,8 +1,8 @@
 package chess.domain.board;
 
-import chess.domain.position.File;
-import chess.domain.position.Square;
-import chess.domain.position.Rank;
+import chess.domain.square.File;
+import chess.domain.square.Square;
+import chess.domain.square.Rank;
 import chess.domain.piece.ColorType;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
