+++ src/main/java/chess/domain/board/dto/MoveCommand.java
@@ -1,6 +1,6 @@
-package chess.dto;
+package chess.domain.board.dto;
 
-import chess.domain.position.Square;
+import chess.domain.square.Square;
 
 public record MoveCommand(Square source, Square destination) {
 }
