+++ src/main/java/chess/domain/square/dto/SquareCreateCommand.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.domain.square.dto;
 
 public class SquareCreateCommand {
