+++ src/main/java/chess/domain/square/dto/SquareDifferent.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.domain.square.dto;
 
 public record SquareDifferent(int rankDiff, int fileDiff) {
 }
