+++ src/main/java/chess/domain/strategy/RookLegalMoveCheckStrategy.java
@@ -1,8 +1,8 @@
 package chess.domain.strategy;
 
 import chess.domain.board.Board;
-import chess.domain.position.Square;
-import chess.dto.SquareDifferent;
+import chess.domain.square.Square;
+import chess.domain.square.dto.SquareDifferent;
 
 public class RookLegalMoveCheckStrategy implements LegalMoveCheckStrategy {
