+++ src/test/java/chess/domain/board/BoardTest.java
@@ -3,10 +3,10 @@ package chess.domain.board;
 import chess.domain.piece.ColorType;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
-import chess.domain.position.File;
-import chess.domain.position.Rank;
-import chess.domain.position.Square;
-import chess.dto.BoardOutput;
+import chess.domain.square.File;
+import chess.domain.square.Rank;
+import chess.domain.square.Square;
+import chess.domain.board.dto.BoardOutput;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
